% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleMessage.r
\name{simpleMessage}
\alias{simpleMessage}
\alias{simpleWarning}
\alias{simpleError}
\alias{simpleDebug}
\title{Create simple log message "object"s.}
\usage{
simpleMessage(message, call = NULL, domain = NULL)

simpleWarning(message, call = NULL, domain = NULL)

simpleError(message, call = NULL, domain = NULL)

simpleDebug(message, call = NULL, domain = NULL)
}
\arguments{
\item{message}{An R object, which is coerced to character (via \code{\link[=as.character]{as.character()}}).}

\item{call}{The calling expression. This option exists for compatibility with the
used in most circumstances. If \code{NULL}, the name of the calling function is
is deduced from the stack frame.}
}
\description{
The following functions serve as log message generators. They will create
S3 objects with, at least, the classes "message" and "condition" (in that
order). These may be used by any of the log messaging functions provided by
this package (e.g. \code{\link[=message]{message()}} or \code{\link[=print.message]{print.message()}}).
}
